Avoid tests hanging when no publicly reachable IP address
On OS X, the public URI selected by Netty is one that
requires a usable network connection (or, at least, an
IP address given to en0). This isn't always the case,
so use the same logic that was used in Selenium to select
the port to use.

We also avoid hardcoding the port to allow more than one
test to be run in parallel.