Add HtmlUnitDriver.setExecutor
Staring with HtmlUnitDriver 2.26 tasks have been submitted asynchronously
in the package private runAsync method. By running the tasks asynchronously
performance improvements can be seen, but the current approach has
limitations:

1) A new Thread is created for every task which is resource intensive
2) There is no way to disable concurrency (i.e. to debug application
issues)
3) ThreadLocal context will be lost without a suitable way to transfer
to the new Thread. This impacts Spring Framework's MockMvc integration
which leverages a ThreadLocal to run a test with a specified user [1].

This commit adds an Executor to HtmlUnitDriver which resolves all of the
issues above. To remain as passive as possible, this commit continues
creating a new Thread per request by default.

[1] https://docs.spring.io/spring-security/site/docs/5.0.x/reference/html/test-mockmvc.html#running-as-a-user-in-spring-mvc-test-with-annotations