Fix HtmlUnitDriver Deadlock
There is a race condition within HtmlUnitDriver.runAsync such
that the wrapped Runnable executes mainCondition.signal() before
mainCondition.awaitUninterruptibly() is invoked. While this can be
reproduced consistently when using an Executor that invokes Runnable.run
on the main Thread, this issue has existed for some time as a race
condition.

This commit ensures that deadlock does not occur by:

1) Only awaitUninterruptibly if the submitted Runnable is still running
2) Ensuring that no signal is sent until after runAsyncRunning is set to
false